feat: make VOM the default and only element detection system
- Remove --vom CLI flag; VOM is now always used for element detection
- Delete old regex-based detection module (12 files, ~800 lines)
- Add VOM-to-Element adapter in session.rs for API compatibility
- Simplify handle_snapshot() to always use VOM path
- Update handle_clear() and handle_select() to use default behavior
- Update lib.rs exports to source Element types from session
- Update tests to reflect VOM-only behavior

This completes the VOM adoption by removing the experimental flag
and making VOM the standard detection mechanism.
