+++ cli/src/daemon/server.rs
@@ -2,9 +2,8 @@ use super::error_messages::{ai_friendly_error, lock_timeout_response};
 use super::lock_helpers::{acquire_session_lock, LOCK_TIMEOUT};
 use super::rpc_types::{Request, Response};
 use super::select_helpers::{navigate_to_option, strip_ansi_codes};
-use crate::detection::{detect_framework, Element, Framework};
 use crate::json_ext::ValueExt;
-use crate::session::SessionManager;
+use crate::session::{Element, SessionManager};
 use crate::sync_utils::mutex_lock_or_recover;
 use crate::wait::{check_condition, StableTracker, WaitCondition};
 use serde_json::{json, Value};
@@ -413,7 +412,6 @@ impl DaemonServer {
         let include_elements = params.bool_or("include_elements", false);
         let interactive_only = params.bool_or("interactive_only", false);
         let compact = params.bool_or("compact", false);
-        let use_vom = params.bool_or("vom", false);
         let req_id = request.id;
 
         self.with_session(&request, session_id, |sess| {
@@ -426,8 +424,8 @@ impl DaemonServer {
             let cursor = sess.cursor();
             let (cols, rows) = sess.size();
 
-            let (elements, stats) = if use_vom {
-                // Use Visual Object Model for element detection
+            let (elements, stats) = if include_elements {
+                // Always use VOM (Visual Object Model) for element detection
                 let vom_components = sess.analyze_screen();
                 let elements_total = vom_components.len();
                 let elements_interactive = vom_components
@@ -461,39 +459,6 @@ impl DaemonServer {
                         "elements_interactive": elements_interactive,
                         "elements_shown": elements_shown,
                         "detection": "vom"
-                    }),
-                )
-            } else if include_elements {
-                let all_elements = sess.detect_elements();
-                let elements_total = all_elements.len();
-                let elements_interactive =
-                    all_elements.iter().filter(|e| e.is_interactive()).count();
-
-                let filtered_elements: Vec<_> = all_elements
-                    .iter()
-                    .filter(|el| {
-                        if interactive_only && !el.is_interactive() {
-                            return false;
-                        }
-                        if compact && !el.is_interactive() && !el.has_content() {
-                            return false;
-                        }
-                        true
-                    })
-                    .map(element_to_json)
-                    .collect();
-
-                let elements_shown = filtered_elements.len();
-
-                (
-                    Some(filtered_elements),
-                    json!({
-                        "lines": screen.lines().count(),
-                        "chars": screen.len(),
-                        "elements_total": elements_total,
-                        "elements_interactive": elements_interactive,
-                        "elements_shown": elements_shown,
-                        "detection": "regex"
                     }),
                 )
             } else {
@@ -1172,14 +1137,8 @@ impl DaemonServer {
             if sess.find_element(element_ref).is_none() {
                 return Response::element_not_found(req_id, element_ref);
             }
-            let screen_text = sess.screen_text();
-            let framework = detect_framework(&screen_text);
-            let result = match framework {
-                Framework::Textual => sess
-                    .pty_write(b"\x01")
-                    .and_then(|_| sess.pty_write(b"\x7f")),
-                _ => sess.pty_write(b"\x15"),
-            };
+            // Send Ctrl+U to clear input (works in most terminals)
+            let result = sess.pty_write(b"\x15");
             if let Err(e) = result {
                 return Response::error(req_id, -32000, &ai_friendly_error(&e.to_string(), None));
             }
@@ -1259,16 +1218,10 @@ impl DaemonServer {
             }
 
             let screen_text = sess.screen_text();
-            let framework = detect_framework(&screen_text);
 
-            let result = match framework {
-                Framework::Unknown => sess
-                    .pty_write(b"\r")
-                    .and_then(|_| sess.pty_write(option.as_bytes()))
-                    .and_then(|_| sess.pty_write(b"\r")),
-                _ => navigate_to_option(sess, &option, &screen_text)
-                    .and_then(|_| sess.pty_write(b"\r")),
-            };
+            // Use keyboard navigation to find and select the option
+            let result =
+                navigate_to_option(sess, &option, &screen_text).and_then(|_| sess.pty_write(b"\r"));
 
             if let Err(e) = result {
                 return Response::error(req_id, -32000, &ai_friendly_error(&e.to_string(), None));
