--- cli/src/detection/frameworks/mod.rs
@@ -1,13 +1,0 @@
-mod bubbletea;
-mod generic;
-mod ink;
-mod inquirer;
-mod ratatui;
-mod textual;
-
-pub use bubbletea::BubbleTeaDetector;
-pub use generic::GenericDetector;
-pub use ink::InkDetector;
-pub use inquirer::InquirerDetector;
-pub use ratatui::RatatuiDetector;
-pub use textual::TextualDetector;
