+++ cli/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod detection;
 pub mod pty;
 pub mod session;
 pub mod sync_utils;
@@ -6,9 +5,10 @@ pub mod terminal;
 pub mod vom;
 pub mod wait;
 
-pub use detection::{Element, ElementDetector, Framework};
 pub use pty::{PtyError, PtyHandle};
-pub use session::{Session, SessionError, SessionId, SessionInfo, SessionManager};
+pub use session::{
+    Element, ElementType, Position, Session, SessionError, SessionId, SessionInfo, SessionManager,
+};
 pub use terminal::{CursorPosition, ScreenBuffer, VirtualTerminal};
 pub use vom::{Component, Role};
 pub use wait::WaitCondition;
