+++ cli/src/main.rs
@@ -4,7 +4,6 @@ mod color;
 mod commands;
 mod daemon;
 mod demo;
-mod detection;
 mod handlers;
 mod json_ext;
 mod pty;
@@ -91,8 +90,7 @@ fn run() -> Result<(), Box<dyn std::error::Error>> {
             interactive_only,
             compact,
             region,
-            vom,
-        } => handlers::handle_snapshot(&mut ctx, elements, interactive_only, compact, region, vom)?,
+        } => handlers::handle_snapshot(&mut ctx, elements, interactive_only, compact, region)?,
         Commands::Click { element_ref } => handlers::handle_click(&mut ctx, element_ref)?,
         Commands::DblClick { element_ref } => handlers::handle_dbl_click(&mut ctx, element_ref)?,
         Commands::Fill { element_ref, value } => {
